+++ android/src/main/java/live/videosdk/webrtc/OrientationAwareScreenCapturer.java
@@ -83,6 +83,11 @@ public class OrientationAwareScreenCapturer implements VideoCapturer, VideoSink 
             throw new RuntimeException("capturer is disposed.");
         }
     }
+
+    public MediaProjection getMediaProjection(){
+        return mediaProjection;
+    }
+
     public synchronized void initialize(final SurfaceTextureHelper surfaceTextureHelper,
                                         final Context applicationContext, final CapturerObserver capturerObserver) {
         checkNotDisposed();
@@ -139,8 +144,8 @@ public class OrientationAwareScreenCapturer implements VideoCapturer, VideoSink 
                 if (mediaProjection != null) {
                     // Unregister the callback before stopping, otherwise the callback recursively
                     // calls this method.
-                    mediaProjection.unregisterCallback(mediaProjectionCallback);
                     mediaProjection.stop();
+//                    mediaProjection.unregisterCallback(mediaProjectionCallback);
                     mediaProjection = null;
                 }
             }