+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+[0.0.11] - 2025.09.15
+- Introduced screen share audio support for Android devices.
+- Updated the Android version.
+- Enabled background camera access on iOS for Picture-in-Picture (PiP) mode.
+- Replaced deprecated `onSurfaceDestroyed` in `SurfaceTextureRenderer` with `onSurfaceCleanup`.
+
 [0.0.10] - 2025.05.23
 - Fixed an issue of deadlock happening when creating a frame cryptor on iOS/macOS.